+++ mrz-3-edition/tester.m
@@ -51,4 +51,4 @@ squaredMultiplicationsTester(2, 3);
 hornerVSAll([10^18, 10 ^ 1, 10^-9], 10^23);  % 9.999999999999999e+63, 9.999999999999999e+63
 parabolaSolverSmartVersion(1, 4, 1);  % -3.7321, -0.2679
 % parabolaSolverTester(10 ^(-3), 0.8, -6 * 10 ^ (-5), -1 * 10 ^ (-5));  % like picture 2.2 in book
-% parabolaSolverRealVersion
+parabolaSolverComplexVersion(9, 1, 1);  % -0.0556 - 0.3287i, -0.0556 + 0.3287i