+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.2.5 - 2017-03-17
+
+### Added
+- parabolaSolverComplexVersion
+
 ## 0.2.4 - 2017-03-16
 
 ### Added